% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_b.R
\name{b_gwas}
\alias{b_gwas}
\title{Perform GWAS for a binary covariate (discovery sample)}
\usage{
b_gwas(plink_path, dis_snp, bp_dis_cov)
}
\arguments{
\item{plink_path}{Character. Path to the PLINK (v2.0) executable.}

\item{dis_snp}{Character. Prefix of PLINK binary files
(.bed/.bim/.fam) for the discovery dataset.}

\item{bp_dis_cov}{Character. File path to phenotype and covariate file.
Columns must be:
\itemize{
\item 1: FID
\item 2: IID
\item 3: Binary phenotype (0/1 or 1/2)
\item 4-K: Covariates
}}
}
\value{
A list containing file paths to GWAS outputs stored in a
temporary directory:
\describe{
\item{glm_file}{Path to PLINK .glm.logistic.hybrid file}
\item{additive_effects}{Path to extracted additive SNP log odds}
\item{tmp_dir}{Temporary directory used by the function}
\item{log_file}{Path to PLINK log file}
\item{n_snps}{Number of SNPs with additive effects}
}
}
\description{
This function runs a PLINK logistic GWAS for a binary phenotype and
extracts additive SNP effects (log odds) for use in downstream
GCIM-GWEIS-Z models. All outputs are saved to a temporary directory.
}
\details{
This function performs the following steps:
\enumerate{
\item Creates a temporary directory for all outputs
\item Runs PLINK logistic GWAS with covariate adjustment
\item Extracts additive SNP effects (TEST equals ADD)
\item Converts odds ratios to log odds for BETA values
\item Saves results to temporary files for pipeline continuity
}
}
