% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2_GCIM_gweis_z.R
\name{ldsc_h2_gcim}
\alias{ldsc_h2_gcim}
\title{Run LDSC heritability estimation (h2) and parse intercept}
\usage{
ldsc_h2_gcim(
  ldsc_path,
  munged_sumstats,
  ref_ld_chr,
  w_ld_chr = ref_ld_chr,
  python = "python",
  tmp_dir = NULL,
  out_prefix = "ldsc_gcim",
  show_log_tail_on_error = TRUE
)
}
\arguments{
\item{ldsc_path}{Character. Path to LDSC directory containing \code{ldsc.py},
OR path directly to \code{ldsc.py}.}

\item{munged_sumstats}{Character OR list. Path to munged \code{.sumstats.gz}, or list
from \code{munge_ldsc_gcim()} containing \verb{$munged_sumstats} or \verb{$files$munged}.}

\item{ref_ld_chr}{Character. Path prefix to reference LD scores (e.g. "eur_w_ld_chr/").}

\item{w_ld_chr}{Character. Path prefix to weights LD scores (default = ref_ld_chr).}

\item{python}{Character. Python executable (default "python").}

\item{tmp_dir}{Character. Optional output directory. If NULL, a new temp directory is created.}

\item{out_prefix}{Character. Output prefix within tmp_dir (default "ldsc_gcim").}

\item{show_log_tail_on_error}{Logical. Print log tail if LDSC fails.}
}
\value{
A list containing LDSC outputs and parsed intercept values.
}
\description{
Run LDSC heritability estimation (h2) and parse intercept
}
