% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2_GCIM_gweis_z.R
\name{ldsc_h2_gcim}
\alias{ldsc_h2_gcim}
\title{Run LDSC heritability estimation (h2) and parse intercept}
\usage{
ldsc_h2_gcim(
  ldsc_path,
  munged_sumstats,
  ref_ld_chr,
  w_ld_chr = ref_ld_chr,
  python = "python",
  tmp_dir = NULL,
  out_prefix = "ldsc_gcim"
)
}
\arguments{
\item{ldsc_path}{Character. Path to LDSC directory containing \code{ldsc.py}.}

\item{munged_sumstats}{Character OR list. Path to munged \code{.sumstats.gz}, or list
from \code{munge_ldsc_gcim()} containing \verb{$munged_sumstats} or \verb{$files$munged}.}

\item{ref_ld_chr}{Character. Path prefix to reference LD scores (e.g. \code{"eur_w_ld_chr/"}).}

\item{w_ld_chr}{Character. Path prefix to weights LD scores (default = \code{ref_ld_chr}).}

\item{python}{Character. Python executable (default \code{"python"}).}

\item{tmp_dir}{Character. Optional output directory. If \code{NULL}, uses a new temp directory.}

\item{out_prefix}{Character. Optional output prefix within \code{tmp_dir} (default \code{"ldsc_gcim"}).}
}
\value{
A list containing LDSC outputs.
}
\description{
Runs \code{ldsc.py --h2} on a munged \code{.sumstats.gz} file and parses the
LDSC intercept (and standard error if present) from the output log.
All outputs are saved to a temporary directory (or a user-provided directory).
}
