% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menge_GCIM_gweis_z.R
\name{munge_ldsc_gcim}
\alias{munge_ldsc_gcim}
\title{Munge interaction summary statistics for LDSC}
\usage{
munge_ldsc_gcim(
  munge_path,
  glm_file,
  interaction_term,
  hm3_snplist,
  chunksize = 1e+05,
  python = "python",
  tmp_dir = NULL,
  out_prefix = "ldsc_gcim"
)
}
\arguments{
\item{munge_path}{Character. Path to \code{munge_sumstats.py}.}

\item{glm_file}{Character OR list. Path to PLINK \verb{.glm.*} file OR list from
\code{q_gweis()}/\code{b_gweis()} containing \verb{$glm_file}.}

\item{interaction_term}{Character. Interaction test name, e.g. \code{"ADDxCOVAR1"}.}

\item{hm3_snplist}{Character. Path to HapMap3 SNP list (\code{w_hm3.snplist}).}

\item{chunksize}{Integer. Chunk size for munging.}

\item{python}{Character. Python executable (default \code{"python"}).}

\item{tmp_dir}{Character. Optional output directory. If \code{NULL}, a new temp directory is created.}

\item{out_prefix}{Character. Optional output prefix within \code{tmp_dir} (default \code{"ldsc_gcim"}).}
}
\value{
A list with munging outputs.
}
\description{
Munge interaction summary statistics for LDSC
}
