% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcim_gweis_q.R
\name{q_gweis}
\alias{q_gweis}
\title{Perform GWEIS for a quantitative phenotype and generate a GCIM input file}
\usage{
q_gweis(
  plink_path,
  tar_mydata,
  tar_pheno_file,
  tar_covar_file,
  int_covar_index = 1,
  out_file = "gcim_prs_int.txt",
  out_prefix = "q_gweis",
  threads = 1
)
}
\arguments{
\item{plink_path}{Character. Path to PLINK2 executable.}

\item{tar_mydata}{Character. Prefix of PLINK binary files (.bed/.bim/.fam).}

\item{tar_pheno_file}{Character. Phenotype file with columns: FID IID PHENO
(PHENO should be in column 3).}

\item{tar_covar_file}{Character. Covariate file with columns: FID IID COVAR1...
(COVAR1 is column 3, used as COVAR1 in PLINK naming).}

\item{int_covar_index}{Integer. Covariate index for interaction term:
1 -> ADDxCOVAR1, 2 -> ADDxCOVAR2, ...}

\item{out_file}{Character. Name of GCIM interaction output file.}

\item{out_prefix}{Character. PLINK output prefix inside tmp_dir (default "q_gweis").}

\item{threads}{Integer. Optional threads for PLINK (default 1).}
}
\value{
A list with file paths and metadata.
}
\description{
Runs PLINK2 linear regression with SNP x covariate interaction and exports
interaction results (ADDxCOVARk) in GCIM format.
}
