% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcim_gweis_q.R
\name{q_gweis}
\alias{q_gweis}
\title{Perform GWEIS for a quantitative phenotype and generate a GCIM input file}
\usage{
q_gweis(
  plink_path,
  tar_mydata,
  tar_pheno_file,
  tar_covar_file,
  int_covar_index = 1,
  out_file = "gcim_prs_int.txt",
  out_prefix = "q_gweis",
  threads = 40
)
}
\arguments{
\item{plink_path}{Character. Path to PLINK2 executable.}

\item{tar_mydata}{Character. Prefix of PLINK binary files (.bed/.bim/.fam).}

\item{tar_pheno_file}{Character OR list. Phenotype file with columns: FID IID PHENO
(PHENO should be in column 3). Can also pass output from previous pipeline step.}

\item{tar_covar_file}{Character OR list. Covariate file with columns: FID IID COVAR1...
(COVAR1 is column 3, should be PRS). Can also pass output from \code{replace_covariate_with_prs()}.}

\item{int_covar_index}{Integer. Covariate index for interaction term:
1 -> ADDxCOVAR1 (typically PRS), 2 -> ADDxCOVAR2, ...}

\item{out_file}{Character. Name of GCIM interaction output file (default "gcim_prs_int.txt").}

\item{out_prefix}{Character. PLINK output prefix inside tmp_dir (default "q_gweis").}

\item{threads}{Integer. Optional threads for PLINK (default 40).}
}
\value{
A list with file paths and metadata.
}
\description{
Runs PLINK2 linear regression with SNP x covariate interaction and exports
interaction results (ADDxCOVARk) in GCIM format.
}
